requirements.txt
Django==4.2.11
django-mptt==0.15.0
Faker==27.0.0
psycopg2-binary==2.9.10
